Moving models to utils
