+++ api-controllers/DashboardController.js
@@ -8,7 +8,7 @@ const Session = require('express-session').Session;
 const config = require('../common/config');
 const errors = require('../common/errors');
 const logger = require('../common/logger');
-const catalog = require('../broker/lib/models').catalog;
+const catalog = require('../common/models').catalog;
 const utils = require('../broker/lib/utils');
 const FabrikBaseController = require('./FabrikBaseController');
 const Forbidden = errors.Forbidden;
