+++ api-controllers/FabrikBaseController.js
@@ -11,7 +11,7 @@ const cf = require('../broker/lib/cf');
 const bosh = require('../data-access-layer/bosh');
 const fabrik = require('../broker/lib/fabrik');
 const backupStore = require('../data-access-layer/iaas').backupStore;
-const catalog = require('../broker/lib/models/catalog');
+const catalog = require('../common/models/catalog');
 const ContinueWithNext = errors.ContinueWithNext;
 const BadRequest = errors.BadRequest;
 const CONST = require('../common/constants');
