+++ api-controllers/ServiceBrokerApiController.js
@@ -5,7 +5,7 @@ const _ = require('lodash');
 const Promise = require('bluebird');
 const errors = require('../common/errors');
 const utils = require('../broker/lib/utils');
-const catalog = require('../broker/lib/models/catalog');
+const catalog = require('../common/models/catalog');
 const FabrikBaseController = require('./FabrikBaseController');
 const lockManager = require('../eventmesh').lockManager;
 const AssertionError = assert.AssertionError;
