+++ api-controllers/ServiceFabrikAdminController.js
@@ -3,7 +3,7 @@
 const _ = require('lodash');
 const Promise = require('bluebird');
 var moment = require('moment-timezone');
-const catalog = require('../broker/lib/models/catalog');
+const catalog = require('../common/models/catalog');
 const errors = require('../common/errors');
 const logger = require('../common/logger');
 const config = require('../common/config');
