+++ broker/lib/fabrik/DBManager.js
@@ -4,7 +4,7 @@ const _ = require('lodash');
 const Promise = require('bluebird');
 const config = require('../../../common/config');
 const logger = require('../../../common/logger');
-const catalog = require('../models/catalog');
+const catalog = require('../../../common/models/catalog');
 const DirectorManager = require('./DirectorManager');
 const bosh = require('../../../data-access-layer/bosh');
 const utils = require('../utils');
