+++ broker/lib/fabrik/DirectorInstance.js
@@ -10,7 +10,7 @@ const logger = require('../../../common/logger');
 const errors = require('../../../common/errors');
 const jwt = require('../jwt');
 const utils = require('../utils');
-const catalog = require('../models').catalog;
+const catalog = require('../../../common/models').catalog;
 const NotFound = errors.NotFound;
 const ServiceInstanceNotFound = errors.ServiceInstanceNotFound;
 const ScheduleManager = require('../jobs');
