+++ broker/lib/fabrik/DirectorTaskPoller.js
@@ -3,7 +3,7 @@
 const bosh = require('../../../data-access-layer/bosh');
 const DirectorManager = require('./DirectorManager');
 const logger = require('../../../common/logger');
-const catalog = require('../models/catalog');
+const catalog = require('../../../common/models/catalog');
 const pubsub = require('pubsub-js');
 const CONST = require('../../../common/constants');
 const config = require('../../../common/config');
