+++ broker/lib/fabrik/DockerInstance.js
@@ -7,7 +7,7 @@ const errors = require('../../../common/errors');
 const utils = require('../utils');
 const docker = require('../docker');
 const BaseInstance = require('./BaseInstance');
-const catalog = require('../models').catalog;
+const catalog = require('../../../common/models').catalog;
 const Timeout = errors.Timeout;
 const ContainerStartError = errors.ContainerStartError;
 const ServiceInstanceAlreadyExists = errors.ServiceInstanceAlreadyExists;
