+++ broker/lib/fabrik/Fabrik.js
@@ -3,7 +3,7 @@
 const assert = require('assert');
 const Promise = require('bluebird');
 const config = require('../../../common/config');
-const catalog = require('../models').catalog;
+const catalog = require('../../../common/models').catalog;
 const Agent = require('./Agent');
 const VirtualHostAgent = require('./VirtualHostAgent');
 const DirectorManager = require('./DirectorManager');
