+++ broker/lib/fabrik/FabrikStatusPoller.js
@@ -3,7 +3,7 @@
 const assert = require('assert');
 const _ = require('lodash');
 const Promise = require('bluebird');
-const catalog = require('../models/catalog');
+const catalog = require('../../../common/models/catalog');
 const pubsub = require('pubsub-js');
 const logger = require('../../../common/logger');
 const ServiceFabrikOperation = require('./ServiceFabrikOperation');
