+++ broker/lib/jobs/BnRStatusPollerJob.js
@@ -11,7 +11,7 @@ const errors = require('../../../common/errors');
 const DeploymentAlreadyLocked = errors.DeploymentAlreadyLocked;
 const config = require('../../../common/config');
 const bosh = require('../../../data-access-layer/bosh');
-const catalog = require('../models').catalog;
+const catalog = require('../../../common/models').catalog;
 const BackupService = require('../../../managers/backup-manager');
 const eventmesh = require('../../../eventmesh');
 const lockManager = eventmesh.lockManager;
