+++ broker/lib/jobs/ServiceInstanceUpdateJob.js
@@ -9,7 +9,7 @@ const errors = require('../../../common/errors');
 const utils = require('../utils');
 const cloudController = require('../cf').cloudController;
 const Fabrik = require('../fabrik');
-const catalog = require('../models/catalog');
+const catalog = require('../../../common/models/catalog');
 const CONST = require('../../../common/constants');
 const ScheduleManager = require('./ScheduleManager');
 const Repository = require('../../../common/db').Repository;
