+++ broker/lib/middleware.js
@@ -8,7 +8,7 @@ const quotaManager = quota.quotaManager;
 const Forbidden = errors.Forbidden;
 const BadRequest = errors.BadRequest;
 const utils = require('../../common/utils');
-const catalog = require('./models/catalog');
+const catalog = require('../../common/models/catalog');
 const lockManager = require('../../eventmesh').lockManager;
 const CONST = require('./../../common/constants');
 const DeploymentAlreadyLocked = errors.DeploymentAlreadyLocked;
