+++ common/models/catalog.js
@@ -1,8 +1,8 @@
 'use strict';
 
 const _ = require('lodash');
-const config = require('../../../common/config');
-const errors = require('../../../common/errors');
+const config = require('../config');
+const errors = require('../errors');
 const Service = require('./Service');
 const ServiceNotFound = errors.ServiceNotFound;
 const ServicePlanNotFound = errors.ServicePlanNotFound;
