+++ data-access-layer/iaas/BackupStore.js
@@ -8,7 +8,7 @@ const errors = require('../../common/errors');
 const NotImplementedBySubclass = errors.NotImplementedBySubclass;
 const logger = require('../../common/logger');
 const CONST = require('../../common/constants');
-const catalog = require('../../broker/lib/models/catalog');
+const catalog = require('../../common/models/catalog');
 const config = require('../../common/config');
 const BaseCloudClient = require('./BaseCloudClient');
 const NotFound = errors.NotFound;
