+++ test/test_broker/EventLogInterceptor.spec.js
@@ -4,7 +4,7 @@ const Promise = require('bluebird');
 const logger = require('../../common/logger');
 const DirectorManager = require('../../broker/lib/fabrik/DirectorManager');
 const DockerManager = require('../../broker/lib/fabrik/DockerManager');
-const catalog = require('../../broker/lib/models/catalog');
+const catalog = require('../../common/models/catalog');
 const utils = require('../../common/utils');
 const config = require('../../common/config');
 const path = require('path');
