+++ test/test_broker/acceptance/dashboard.docker.spec.js
@@ -4,7 +4,7 @@ const _ = require('lodash');
 const parseUrl = require('url').parse;
 const lib = require('../../../broker/lib');
 const app = require('../support/apps').external;
-const catalog = lib.models.catalog;
+const catalog = require('../../../common/models').catalog;
 const docker = lib.docker;
 const fabrik = lib.fabrik;
