+++ test/test_broker/acceptance/service-broker-api-2.0.instances.director.spec.js
@@ -7,7 +7,7 @@ const Promise = require('bluebird');
 const app = require('../support/apps').internal;
 const utils = lib.utils;
 const config = require('../../../common/config');
-const catalog = lib.models.catalog;
+const catalog = require('../../../common/models').catalog;
 const fabrik = lib.fabrik;
 const ScheduleManager = require('../../../broker/lib/jobs');
 const CONST = require('../../../common/constants');
