+++ test/test_broker/acceptance/service-broker-api.instances.director.spec.js
@@ -5,7 +5,7 @@ const lib = require('../../../broker/lib');
 const Promise = require('bluebird');
 const app = require('../support/apps').internal;
 const config = require('../../../common/config');
-const catalog = lib.models.catalog;
+const catalog = require('../../../common/models').catalog;
 const fabrik = lib.fabrik;
 const iaas = require('../../../data-access-layer/iaas');
 const backupStore = iaas.backupStore;
