+++ test/test_broker/acceptance/service-broker-api.instances.docker.spec.js
@@ -4,7 +4,7 @@ const _ = require('lodash');
 const parseUrl = require('url').parse;
 const lib = require('../../../broker/lib');
 const app = require('../support/apps').internal;
-const catalog = lib.models.catalog;
+const catalog = require('../../../common/models').catalog;
 const docker = lib.docker;
 const config = require('../../../common/config');
 const fabrik = lib.fabrik;
