+++ test/test_broker/acceptance/service-fabrik-api-2.0.instances.director.spec.js
@@ -7,7 +7,7 @@ const lib = require('../../../broker/lib');
 const ScheduleManager = require('../../../broker/lib/jobs');
 const CONST = require('../../../common/constants');
 const apps = require('../support/apps');
-const catalog = lib.models.catalog;
+const catalog = require('../../../common/models').catalog;
 const config = require('../../../common/config');
 const errors = require('../../../common/errors');
 const fabrik = lib.fabrik;
