+++ test/test_broker/fabrik.DockerManager.spec.js
@@ -4,7 +4,7 @@ const lib = require('../../broker/lib');
 const fabrik = require('../../broker/lib/fabrik/Fabrik');
 const config = require('../../common/config');
 const portRegistry = lib.docker.portRegistry;
-const catalog = lib.models.catalog;
+const catalog = require('../../common/models').catalog;
 const DockerManager = lib.fabrik.DockerManager;
 
 describe('fabrik', function () {
