+++ test/test_broker/jobs.ServiceInstanceUpdateJob.spec.js
@@ -2,7 +2,7 @@
 const _ = require('lodash');
 const Promise = require('bluebird');
 const CONST = require('../../common/constants');
-const catalog = require('../../broker/lib/models/catalog');
+const catalog = require('../../common/models/catalog');
 const utils = require('../../broker/lib/utils');
 const config = require('../../common/config');
 const errors = require('../../common/errors');
