+++ test/test_broker/mocks/uaa.js
@@ -4,9 +4,8 @@ const _ = require('lodash');
 const jwt = require('jsonwebtoken');
 const nock = require('nock');
 const parseUrl = require('url').parse;
-const lib = require('../../../broker/lib');
 const config = require('../../../common/config');
-const catalog = lib.models.catalog;
+const catalog = require('../../../common/models').catalog;
 const tokenEndpointUrl = 'https://uaa.bosh-lite.com';
 const authorizationEndpointUrl = 'https://login.bosh-lite.com';
 const jwtToken = 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjI1MzQwMjMwMDc5OSwidXNlcl9pZCI6Im1lIiwic2NvcGUiOlsib3BlbmlkIiwiY2xvdWRfY29udHJvbGxlci53cml0ZSIsImNsb3VkX2NvbnRyb2xsZXJfc2VydmljZV9wZXJtaXNzaW9ucy5yZWFkIl19.ClDfNqT9T1_5LicTpqNrHJ9Fv-UwkLVZNWG71PjCAVQ';
