+++ test/test_broker/quota.QuotaManager.spec.js
@@ -5,7 +5,7 @@ const Promise = require('bluebird');
 const quota = require('../../broker/lib/quota');
 const proxyquire = require('proxyquire');
 const quotaManager = quota.quotaManager;
-const catalog = require('../../broker/lib/models/catalog');
+const catalog = require('../../common/models/catalog');
 const CloudControllerClient = require('../../broker/lib/cf/CloudControllerClient');
